+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version TBD
+## Bug Fixes
+* Fixed incorrect Android package version selection when a mix of
+  version-locked and non-version-locked packages are specified.
+* Fixed non-deterministic Android package version selection to select
+  the highest version of a specified package rather than the last
+  package specification passed to the Gradle resolution script.
+
 # Version 1.2.74 - Jun 19, 2018
 ## New Features
 * Added workaround for broken AndroidManifest.xml variable replacement in